+++ src/main/java/com/project/likelion14thbe/domain/order/exception/OrderErrorCode.java
@@ -1,4 +1,4 @@
-package com.project.likelion14thbe.domain.order.execption;
+package com.project.likelion14thbe.domain.order.exception;
 
 import com.project.likelion14thbe.global.apiPayload.code.BaseErrorCode;
 import lombok.AllArgsConstructor;
